Reworked Tile Connection + Fix Upgrade Conveyor item not showing correctly

+++ Simple Game Engine/OreTile.h
@@ -6,7 +6,7 @@ class OreTile : public Tile
 public:
 
 	OreTile() :Tile(), drop(Item::Iron) {};
-	OreTile(std::string tname,float tID, Item i,float d, float rot, bool cRot,bool rRot, PlaceableOn pOn) : Tile(tname,tID, rot, cRot,rRot, Type::Ore, pOn),drop(i),density(d) {};
+	OreTile(std::string tname,int tID, Item i,float d, float rot, bool cRot,bool rRot, PlaceableOn pOn) : Tile(tname,tID, rot, cRot,rRot, Type::Ore, pOn),drop(i),density(d) {};
 
 	Item drop;
 	float density;
